Some minor improvements on the fix
1. Use expand-file-name to deal with the case where the user deletes
the initial / or ~/ and the minibuffer therefore contains only a
relative path.

2. Call completion-boundaries not on the entire minibuffer-contents,
but only the part before point, specifying the part after point as a
suffix. This is to get the correct count of components to replace when
the user is completing a component that is not the last one.